Move layout code into new Layout class
