+++ lib/flutterby/view.rb
@@ -1,4 +1,5 @@
 require 'benchmark'
+require 'flutterby/layout'
 
 module Flutterby
   class View
@@ -26,7 +27,7 @@ module Flutterby
 
         # Apply layouts
         if opts[:layout] && node.page?
-          @_body = apply_layout!(@_body)
+          Layout.apply!(self)
         end
       end
 
@@ -42,44 +43,6 @@ module Flutterby
       logger.debug "Rendered #{node.url.colorize(:blue)} in #{sprintf("%.1fms", time * 1000).colorize(color)}"
 
       @_body
-    end
-
-    def apply_layout!(input)
-      collect_layouts.inject(input) do |output, layout|
-        tilt = Flutterby::Filters.tilt(layout.ext, layout.source)
-        tilt.render(self) { output }.html_safe
-      end
-    end
-
-    def collect_layouts
-      layouts = []
-
-      # Collect layouts explicitly configured for node
-      if defined? node.layout
-        Array(node.layout).each do |sel|
-          # If a false is explicity specified, that's all the layouts
-          # we're expected to render
-          return layouts if sel == false
-
-          if layout = node.find(sel)
-            layouts << layout
-          else
-            raise "No layout found for path expression '#{sel}'"
-          end
-        end
-      end
-
-      # Decide on a starting node for walking the tree upwards
-      start = layouts.any? ? layouts.last.parent : node
-
-      # Walk the tree up, collecting any layout files found on our way
-      TreeWalker.walk_up(start) do |node|
-        if layout = node.sibling("_layout")
-          layouts << layout
-        end
-      end
-
-      layouts
     end
 
     def to_s
